scripts/requirements.txt
numpy>=1.20
Pillow>=9.0
opencv-python>=4.5
